chain/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-allauth==0.41.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-import-export==2.0
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
pbr==5.4.4
Pillow==7.0.0
protobuf==3.11.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
stevedore==1.31.0
tablib==0.14.0
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
xlrd==1.2.0
xlwt==1.3.0
